pyproject.toml
[project]
name = "cbb-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=6.29.5",
    "lxml>=5.3.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.28.0",
]
